+++ src/Model/src/Attitude.cpp
@@ -76,8 +76,8 @@ const AttitudeData &Attitude::interpolate(std::int64_t timestamp, TimeVariableDa
             m_currentData.velocityBodyY = SkyMath::interpolateLinear(p1->velocityBodyY, p2->velocityBodyY, tn);
             m_currentData.velocityBodyZ = SkyMath::interpolateLinear(p1->velocityBodyZ, p2->velocityBodyZ, tn);
 
-            // On ground
-            m_currentData.onGround = SkyMath::interpolateNearestNeighbour(p1->onGround, p2->onGround, tn);
+            // On ground (boolean value - no interpolation)
+            m_currentData.onGround = p1->onGround;
 
             m_currentData.timestamp = adjustedTimestamp;
         } else {