+++ src/Plugins/Connect/PathCreator/src/PathCreatorPlugin.h
@@ -64,8 +64,6 @@ protected:
     ConnectPluginBaseSettings &getPluginSettings() const noexcept override;
     std::optional<std::unique_ptr<OptionWidgetIntf>> createExtendedOptionWidget() const noexcept override;
 
-    bool isTimerBasedRecording(SampleRate::SampleRate sampleRate) const noexcept override;
-    
     bool onSetupFlightSimulatorShortcuts() noexcept override;
 
     bool onInitialPositionSetup(const InitialPosition &initialPosition) noexcept override;
@@ -74,7 +72,7 @@ protected:
 
     bool onStartFlightRecording() noexcept override;
     bool onStartAircraftRecording() noexcept override;
-    void onRecordingPaused(Initiator initiator, bool paused) noexcept override;
+    void onRecordingPaused(Initiator initiator, bool enable) noexcept override;
     void onStopRecording() noexcept override;
 
     bool onStartReplay(std::int64_t currentTimestamp) noexcept override;
@@ -82,7 +80,6 @@ protected:
     void onStopReplay() noexcept override;
 
     void onSeek(std::int64_t currentTimestampS, SeekMode seekMode) noexcept override;
-    void onRecordingSampleRateChanged(SampleRate::SampleRate sampleRate) noexcept override;
 
     bool sendAircraftData(std::int64_t currentTimestamp, TimeVariableData::Access access, AircraftSelection aircraftSelection) noexcept override;
     bool isConnectedWithSim() const noexcept override;
@@ -95,9 +92,6 @@ protected:
 
     bool onRequestLocation() noexcept override;
     bool onRequestSimulationRate() noexcept override;
-
-protected slots:
-    void recordData() noexcept override;
 
 private:
     const std::unique_ptr<PathCreatorPluginPrivate> d;
@@ -117,6 +111,7 @@ private:
 
 private slots:
     void replay() noexcept;
+    void recordData() noexcept;
 };
 
 #endif // PATHCREATORPLUGIN_H